refactor | htmlElements: mv into src dir
