+++ src/pages/_app.tsx
@@ -12,7 +12,7 @@ import '@/styles/reset.css';
 import '@/styles/global.scss';
 
 import * as gtag from '@/src/helpers/googleAnalytics';
-import HTMLHead from '@/components/htmlElements/HTMLHead';
+import HTMLHead from '@/src/htmlElements/HTMLHead';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export default function App({ Component, pageProps }: any) {
