+++ src/pages/index.tsx
@@ -2,7 +2,7 @@
 // All rights reserved.
 import React from 'react';
 
-import HTMLBody from '@/components/htmlElements/HTMLBody';
+import HTMLBody from '@/src/htmlElements/HTMLBody';
 import Home from './home';
 
 export default function PageRoot() {
